feat: Add support for wildcard address objects
Added ip-wildcard as a type in the AddressObject PanObject

PR #228
Closes #226 

I chose not to use a version profile for this parameter because the user will get an error on pre-9.0 versions either way. For example, there's no reason to omit the ip-wildcard on 8.1 because that AddressObject would be invalid anyway.