chore: Rebrand to pan-os-python
This is a major change, but should be considered a breaking change,
though it most likely doesn't break anything

The name pandevice is confusing and awful. It originally came from the first
class in the library which was class PanDevice(object): but since the library
has grown by orders of magnitude since then, this name is no longer
descriptive or applicable.