ci: Disable Windows tests
Something is suddenly wrong with installing dependencies in Windows. I
can't figure out what because this seems like a change in the GitHub
Actions runner, not in the actions or our code. Further troubleshooting
is needed, but in the meantime I will temporarily disable Windows as a
test environment to get things running again.