ci: Removing py3.5
We still support python 3.5 in this version, but it is becoming more
difficult as ci tests are starting to fail and be hard to maintain. Rather
than continue to fight to support py3.5, we will remove py3.5 support
in the next major release. For now, we are just removing it from CI
tests.