feat: Add device group hierarchy support
This adds knowledge and handling of device group hierarchy to
pan-os-python.

Handling of this is moved to the `opstate` namespace since device
group hierarchy is not stored in the XPATH of PAN-OS.

Note that the "async" job to update a device group's parent seems
to finish instantaneously.  Most times in my testing the job has
already finished by the time I could query the job status with
pan-os-python.