fix: Correcting profile xpaths
This fixes #266.  The correct XPATH for profiles when directly configured
in Panorama is /config/panorama not /config/shared.  To support this, a new
object ROOT was created: PANORAMA_VSYS, which allows a fallback ROOT of VSYS,
but will otherwise be /config/panorama if the object is directly connected to
a Panorama PanDevice object.