fix: Fixes `refresh()` for attrib style params
This fixes "attrib" style params, which are (currently) only present in
the various rule objects in policies.py.

Doing `rule.apply()` when the uuid is set and the name is unchanged
does not cause PAN-OS to complain, so this fix does not cause any
regressions / change in standard behavior.

Doing `rule.apply() when the uuid is set and the name IS changed
causes PAN-OS to silently create another rule with a new uuid.  This
of course means that the uuid of the object that was applied differs
between the user's object hierarchy and what's on live.

In light of this, we might want to investigate how to handle changing the
unique identifier from the name to the uuid of the rule, but some research
is needed on if this is actually possible from an XPATH perspective as well
as how to gracefully handle backwards compatibility (before the uuid field existed).

Fixes #392