[ISSUE-441]  Modified for loop to be less hokey and in-line with contribution standards
Problem?

Breaking up the defaults for loop ended up making the overall code more
complicated than it needed to be, and made it harder to evaluate the
hip-profiles change on its own.

Solution!

Simplified the code by adding a second loop after the defaults loop to
add the additional profile just to the `hip-profiles` path.