feat: Software updater path can exclude base image
The software updater installed the base image before installing the
patch image for each major and minor release. This comes with the
disadvantage that the base image might have bugs and is running for a
short time before the patch upgrade begins. To avoid these potential
bugs in the base image, a new option to skip installing the base image
is now possible.